1Week

+++ src/main/java/com/example/mission_kimeunji/gramgram/boundedContext/likeablePerson/repository/LikeablePersonRepository.java
@@ -3,10 +3,9 @@ package com.example.mission_kimeunji.gramgram.boundedContext.likeablePerson.repo
 
 import com.example.mission_kimeunji.gramgram.boundedContext.likeablePerson.entity.LikeablePerson;
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
 
 import java.util.List;
 
-public interface LikeablePersonRepository extends JpaRepository<LikeablePerson, Integer> {
+public interface LikeablePersonRepository extends JpaRepository<LikeablePerson, Long> {
     List<LikeablePerson> findByFromInstaMemberId(Long fromInstaMemberId);
 }
